First version of parallel testing
Adds ./para which will manage exclusive access to a databae
Adds ./client which a temporary test client to validate that
the named pipe is held exclusively between client and para.
This way if a 2nd client instance would try to read the named pipe (at
'/tmp/db', it would block until the 1st client would be done with it.