Fix to delete Role row when a Manager is deleted.
If one deletes a project, the delete cascades to Manager.
But it would not cascade to Role, because there is not relationship or
foreign key in the that direction (i.e. the Manager depends on the Role,
but the Role does not depend on the Manager).
This is a by-product of how polymorphism is implemented (in the simplest
case) with sqlalchemy.
So, in order to delete the Role row, we have to explicitly load the
Roles in Project.managers and set the relationship to
'passive_delete=False', which forces sqlalchemy to load and delete each
Manager instance, instead of relying on the database via a CASCADE
event.