Add a more robust dependency detection in cache
If an object being invalidated is not present in the cache
we must verify if any related object is in the cache and if so,
invalidate that one.
The previous algorithm would fail if the attribute name for a related
object was not matching its table name.
For instance, Work.offer as opposed to Work.work_offer.
Reading more SQLAlchemy documentation, I found that the right way to
do this is to check the relationship objects of the model themselves.