Fix for Work state machine when resolving a mediation.
The previous state model would transition from
'in_mediation' to 'resolved'
But the 'resolved' state when written to handle exiting
a 'blocked' state.
The bug was 2-fold:
1/ the event did not have a the proper state function signature,
which requires a comment. That would create a 500 error when
client and contractor would reach an agreement at the end of a
mediation.

2/ it would overwrite the last blockage timestamp.

The fix routes 'in_mediation' to 'in_progress' with a 'resolve' event.

Also, it moves the offer foreign key to the work model, as the work
model is unconditionally related to an offer.
This last change will require running:

api $ . env/docker.bash
api $ _repopulate --hard