Fix for re-raised ValidationErrors
The context manager would catch the marshmallow ValidationError (line
66)
Re-raise (line 67) as a rebase ValidationError
then catch it again, this time as a generic Exception (line 70)
and finally re-raise it as a ServerError.
That behavior would by-pass the try-catch in the auth resource and
return a 500 error instead of the proper 401 HTTP code.