Fix for role switching logic
The ratings calculation introduced a side-effect whereby
we set the current_role in order to query some Reviews,
but we never set the current role back to original value
before the query.
The proper way to do the query was instead to temporarily
change the filter_based_on_current_role class attribute of Review.