Remove 'technologies' field from SkillSet model
The shape of the data for technologies will change a lot in the coming
weeks. Instead of storing it in the databae and having to do lots
of database migrations, let's store the data on disk, pickled.
We can then provide a 'view' of the data suitable for the UI client,
as a derived attribute in the model.