Add a FIFO + subprocess server protocol
This allows us to write parsers in any language.
To make a new parser server, one only needs to implement
'on_new_input' and return a computed object, and 'on_error',
in case any exception is thrown inside 'on_new_input'.

A full example for the Python 2 interpreter is given.