Remove retry fn from GithubScanner
Instead we will fix the real underlying problem:
the volume 'repos' is shared between rq_default
containers, which I suspect causes the intermitent
errors when 2 containers try to write at the same time.

There is no need for a durable storage for repos
since 99% of repos will be scanned within minutes and
we delete them as we go, to keep storage needs constant.
Instead we should mount a RAM-based volume with enough capacity.
1GB should be plenty.

That way we remove a point of resource contention, making the
RQ workers run faster and reliably.

We should also stop writing the results of scanning into /crawler,
for the same reasons.
Instead we should use S3 (more parallel, no contention, simpler
design layout, cheaper to run).