Massive refactor of the skills scanner design.
We had 3 different clients for the different languages (JS, Java,
Python).
This refactoring simplifies the design by settling on a single
client class for remote scanner clients.

I also classes TechonologyScanner and ScannerClient abstract
base classes.

Python uses a ScannerClient implementation that relies on
a subprocess with pipes, while JS and Java use a RemoteScannerClient
to communicate over TCP.
Ultimately I'll change PythonClient to be like JS & Java and be
a remote TCP service available as a couple of independent Docker services:
one for Python 2 and the other for Python 3.