Add back parallel cloning and make cloning more reliable
cloning for crawling no longer relies on named Docker volumes.
Instead the rq_default container either clones in RAM or or
on the default allocated disk for the container.
Thus no sharing between containers and now we can re-enable
multiple crawlers working in parallel.

If a repo does not fit in RAM, we clone on disk.